Modify createDivs() function
Changed the name of createDivs() function to createGrid() because this
function will be used when a button is clicked to generate a new grid.

Also added a parameter to the function, squarePerSide that represents
the number of squares that the new grid should have and finally changed
the for condition to reflect that.