Create deleteContainer and updateGrid functions
The deleteContainer() function is a function that deletes the <div>
that contains the grid; it's used when the updateGrid() function is
executed.

updateGrid() Function is the function that runs when the button is
clicked. This function will delete the existing grid when the user
wants a new one, then creates a new div for the grid, then ask the
user for the grid's new size and finally create it.