Set a random color to a <div> when hovered
Created a function to set a random color in a <div> when hovered. Also
set a condition to just apply one random color instead of a new one
everytime the mouse hovers a <div>.

To track the times a <div> is hovered, it is set this number in the
textContent of the <div>, but the font size is set to 0 to not display
that number to the user.

For now, the <div> is turned to black color when the mouse has over one
<div> ten times; the idea is to just add 10% black for each time the
mouse hovers that same element.