Add ten percent black to a square when hover
To do this first we extract the rgb value from the backgroundColor
attribute of the <div>; there is a function called getRGBValues() that
does this. Before returning the values, it is used another function,
convertToInts() because in the getRGBValues() it's values are strings.

Then to slowly turn to black the square's color we use the
reduceTenPercent() function. This function receives the rgb values and
substracts 10 percent to each one and return a new array. And finally
we apply those values to set the new backgroundColor.